Implement first pass at type-hinting
This isn't pretty yet, but this works for the CAMS use cases. This allows
a configuration hash to be passed in that, for now, just includes type
information. That lets us short-circuit typecasting for filtering so that
we don't have to hit the database to find the types of computed attributes.